+++ tests/e2e/specs/wp-admin/test-add-shipping-zones.js
@@ -1,0 +1,6 @@
+/*
+ * Internal dependencies
+ */
+const { runAddNewShippingZoneTest } = require( '@woocommerce/e2e-core-tests' );
+
+runAddNewShippingZoneTest();